Made a little more user friendly

+++ Program.cs
@@ -69,9 +69,9 @@ public record FileOutput
 static class Api
 {
     public static Realm Realm { get; private set; } = null!;
-    public static string LazerPath { get; private set; } = null!;
+    private static string LazerPath { get; set; } = null!;
 
-    public static string OutPath { get; private set; } = null!;
+    private static string OutPath { get; set; } = null!;
 
     public static void Init(string? _lazerPath, string outPath)
     {
@@ -156,14 +156,17 @@ static class Api
 
     public static void ValidatePaths()
     {
+        Console.WriteLine("Validating paths, this can take a while...");
+
         var files = Directory.EnumerateFiles(OutPath, "*", SearchOption.AllDirectories);
 
         foreach (var file in files)
         {
-            var info = System.IO.File.ResolveLinkTarget(file, false)!;
-
-            if (!info.Exists)
+            var info = System.IO.File.ResolveLinkTarget(file, false);
+            
+            if (info is not null && !info.Exists)
             {
+                Console.WriteLine($"Removing {file}");
                 System.IO.File.Delete(file);
             }
         }
@@ -174,9 +177,12 @@ static class Api
         {
             if (!Directory.EnumerateFileSystemEntries(dir).Any())
             {
+                Console.WriteLine($"Removing {dir}");
                 Directory.Delete(dir);
             }
         }
+
+        Console.WriteLine("Validated paths");
     }
 
     public static void CreateLinksAll(bool isCopy = false)
@@ -218,14 +224,13 @@ static class Api
                 System.IO.File.Copy(src, dst, false);
                 return;
             }
-            
+
             System.IO.File.CreateSymbolicLink(dst, src);
         }
     }
 }
 
 [Preserve(AllMembers = true)]
-
 internal static class Program
 {
     [Preserve(AllMembers = true)]
@@ -235,21 +240,44 @@ internal static class Program
         public required string OutPath { get; init; }
         public string? ReplayPath { get; init; }
         public bool IsCopy { get; init; }
+        public bool IsQuiet { get; init; }
+        public bool IsVerbose { get; init; }
         public bool All { get; init; }
         public bool Validate { get; init; }
         public string? MD5Hash { get; init; }
-        public long OnlineID { get; init; }
+        public long? OnlineID { get; init; }
+
+        public bool CannotRun()
+        {
+            return this is { All: false, Validate: false, MD5Hash: null, OnlineID: null, ReplayPath: null };
+        }
+        
+        public bool CannotContinue()
+        {
+            return this is { All: false, MD5Hash: null, OnlineID: null, ReplayPath: null };
+        }
     }
-    
+
     private static void Run(Args args)
     {
+        if (args.CannotRun())
+        {
+            Console.WriteLine("Nothing to do, use one of [-a, -m, -i, -r] options");
+            return;
+        }
+
         Api.Init(args.LazerPath, args.OutPath);
+
+        if (args.IsQuiet)
+        {
+            Console.SetOut(TextWriter.Null);
+        }
 
         if (args.Validate)
         {
             Api.ValidatePaths();
 
-            if (args.OnlineID == 0 || args.MD5Hash is null || args.ReplayPath is null)
+            if (args.CannotContinue())
             {
                 return;
             }
@@ -258,23 +286,51 @@ internal static class Program
         if (args.All)
         {
             Api.CreateLinksAll(args.IsCopy);
-            return;
         }
 
-        if (args.OnlineID == 0)
+        if (args.MD5Hash is not null || args.ReplayPath is not null)
         {
             var md5hash = args.MD5Hash ?? Api.GetMD5HashFromReplay(args.ReplayPath);
 
             var beatmap = Api.Realm.All<Beatmap>().First(b => b.MD5Hash == md5hash);
 
             Api.CreateLinks(beatmap, args.IsCopy);
+
+            return;
         }
-        else
+
+        if (args.OnlineID == 0)
         {
             var beatmap = Api.Realm.All<Beatmap>().First(b => b.OnlineID == args.OnlineID);
 
             Api.CreateLinks(beatmap, args.IsCopy);
         }
+    }
+
+    private static bool CheckDrives(string lazerPath, string outPath)
+    {
+        if (Path.GetPathRoot(lazerPath) != Path.GetPathRoot(outPath))
+        {
+            Console.WriteLine("These are on different drives");
+            Console.WriteLine(" ");
+            Console.WriteLine($"Lazer path: {lazerPath}");
+            Console.WriteLine($"Output Path: {outPath}");
+            Console.WriteLine(" ");
+            return true;
+        }
+
+        return false;
+    }
+
+    private static bool ReadKey()
+    {
+        var key = Console.ReadKey();
+        return key.Key == ConsoleKey.Escape;
+    }
+
+    private static bool CheckArg(string arg, IEnumerable<Option> options)
+    {
+        return options.Any(o => o.Name == arg || o.Aliases.Any(a => a == arg));
     }
 
     public static void Main(string[] _args)
@@ -283,35 +339,53 @@ internal static class Program
 
         Option<DirectoryInfo> lazerPath = new("--dir", "-d")
         {
+            DefaultValueFactory = _ => new DirectoryInfo(Api.GetDefaultLazerPath()),
             Description = "Path to lazer directory"
-        };
-        Option<FileInfo> replayPath = new("--replay", "-r")
-        {
-            Description = "Path to replay file"
         };
         Option<DirectoryInfo> outPath = new("--out", "-o")
         {
-            Required = true,
+            DefaultValueFactory = _ =>
+                new DirectoryInfo("./YOU-CAN-RENAME-THIS-AND-MOVE-THIS-ON-SAME-DRIVE"),
             Description = "Path to output directory"
+        };
+        Option<FileInfo?> replayPath = new("--replay", "-r")
+        {
+            DefaultValueFactory = _ => null,
+            Description = "Path to replay file"
         };
         Option<bool> isCopy = new("--copy", "-c")
         {
+            DefaultValueFactory = _ => false,
             Description = "copy instead of symlinks"
+        };
+        Option<bool> isQuiet = new("--quiet", "-q")
+        {
+            DefaultValueFactory = _ => false,
+            Description = "suppress output"
+        };
+        Option<bool> isVerbose = new("--verbose")
+        {
+            DefaultValueFactory = _ => false,
+            Description = "show detailed info (does nothing now)"
         };
         Option<bool> all = new("--all", "-a")
         {
-            Description = "symlink all beatmaps"
+            DefaultValueFactory = _ => false,
+            Description = "symlink all beatmaps, enabled if ran with no other args"
         };
         Option<bool> validate = new("--validate", "-v")
         {
+            DefaultValueFactory = _ => false,
             Description = "validate symlinks in output directory"
         };
-        Option<string> md5hash = new("--md5", "-m")
+        Option<string?> md5hash = new("--md5", "-m")
         {
+            DefaultValueFactory = _ => null,
             Description = "beatmap md5hash"
         };
-        Option<long> onlineId = new("--id", "-i")
+        Option<long?> onlineId = new("--id", "-i")
         {
+            DefaultValueFactory = _ => null,
             Description = "beatmap online id (not beatset)"
         };
 
@@ -319,19 +393,136 @@ internal static class Program
         root.Options.Add(outPath);
         root.Options.Add(replayPath);
         root.Options.Add(isCopy);
+        root.Options.Add(isQuiet);
+        root.Options.Add(isVerbose);
         root.Options.Add(all);
         root.Options.Add(validate);
         root.Options.Add(md5hash);
         root.Options.Add(onlineId);
 
+        var defLazerPath = lazerPath.DefaultValueFactory(null!).FullName;
+        var defOutPath = outPath.DefaultValueFactory(null!).FullName;
+
+        if (!Directory.Exists(defLazerPath))
+        {
+            Console.WriteLine($"{defLazerPath} not found");
+            Console.WriteLine("Maybe you have a custom lazer install? use -d to specify where to find it");
+            return;
+        }
+
+        switch (_args.Length)
+        {
+            // drag and dropped on exe
+            case 1:
+            {
+                if (CheckArg(_args[0], root.Options))
+                {
+                    break;
+                }
+
+                var args = new Args
+                {
+                    LazerPath = defLazerPath,
+                    OutPath = Path.GetFullPath(_args[0]),
+                    ReplayPath = null,
+                    IsCopy = false,
+                    IsQuiet = false,
+                    IsVerbose = true,
+                    All = true,
+                    Validate = true,
+                    MD5Hash = null,
+                    OnlineID = null,
+                };
+
+                if (System.IO.File.Exists(args.OutPath))
+                {
+                    Console.WriteLine($"{args.OutPath} is not a directory");
+                    ReadKey();
+                    return;
+                }
+
+                if (CheckDrives(args.LazerPath, args.OutPath))
+                {
+                    return;
+                }
+
+                if (Directory.Exists(args.OutPath) && Directory.EnumerateFileSystemEntries(args.OutPath).Any())
+                {
+                    Console.ForegroundColor = ConsoleColor.Yellow;
+                    Console.WriteLine("");
+                    Console.WriteLine("Output directory contains files, confirm its correct");
+                    Console.ResetColor();
+                    Console.WriteLine("");
+                }
+
+                Console.WriteLine("This will validate existing symlinks and symlink all beatmaps!");
+                Console.WriteLine(" ");
+                Console.WriteLine($"Lazer path: {args.LazerPath}");
+                Console.WriteLine($"Output Path: {args.OutPath}");
+                Console.WriteLine(" ");
+                Console.WriteLine("Press <ESC> or <Ctrl+C> to exit...");
+                Console.WriteLine("Press any key to continue...");
+
+                if (ReadKey())
+                {
+                    return;
+                }
+
+                Run(args);
+
+                return;
+            }
+            // double-clicking exe directly
+            case 0:
+            {
+                var args = new Args
+                {
+                    LazerPath = defLazerPath,
+                    OutPath = defOutPath,
+                    ReplayPath = null,
+                    IsCopy = false,
+                    IsQuiet = false,
+                    IsVerbose = true,
+                    All = true,
+                    Validate = false,
+                    MD5Hash = null,
+                    OnlineID = null,
+                };
+
+                if (CheckDrives(args.LazerPath, args.OutPath))
+                {
+                    return;
+                }
+
+                Console.WriteLine("This will symlink all beatmaps!");
+                Console.WriteLine(" ");
+                Console.WriteLine($"Lazer path: {args.LazerPath}");
+                Console.WriteLine($"Output Path: {args.OutPath}");
+                Console.WriteLine(" ");
+                Console.WriteLine("Press <ESC> or <Ctrl+C> to exit...");
+                Console.WriteLine("Press any key to continue...");
+
+                if (ReadKey())
+                {
+                    return;
+                }
+
+                Run(args);
+
+                return;
+            }
+        }
+
         root.SetAction(parsed =>
         {
             var parsedArgs = new Args
             {
-                LazerPath = parsed.GetValue(lazerPath)?.FullName ?? Api.GetDefaultLazerPath(),
-                OutPath = parsed.GetRequiredValue(outPath).FullName,
+                LazerPath = parsed.GetValue(lazerPath)?.FullName!,
+                OutPath = parsed.GetValue(outPath)?.FullName!,
                 ReplayPath = parsed.GetValue(replayPath)?.FullName,
                 IsCopy = parsed.GetValue(isCopy),
+                IsQuiet = parsed.GetValue(isQuiet),
+                IsVerbose = parsed.GetValue(isVerbose),
                 All = parsed.GetValue(all),
                 Validate = parsed.GetValue(validate),
                 MD5Hash = parsed.GetValue(md5hash),
